requirements.txt
einops==0.6.0
imageio==2.16.1
matplotlib==3.5.1
numpy==1.21.2
Pillow==9.4.0
pytorch_lightning==1.5.10
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0